+++ app/src/androidTest/java/woowacourse/movie/MainActivityTest.kt
@@ -7,6 +7,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Rule
 import org.junit.Test
+import woowacourse.movie.MovieList.MainActivity
 
 class MainActivityTest {
     @get:Rule
