+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".BookingActivity"
+            android:name=".booking.BookingActivity"
             android:exported="false"/>
         <activity
-            android:name=".BookingResultActivity"
+            android:name=".BookingResult.BookingResultActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".MovieList.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
