+++ app/src/main/java/woowacourse/movie/MovieList/MovieListAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.adapter
+package woowacourse.movie.MovieList
 
 import android.content.Context
 import android.view.LayoutInflater
