+++ app/src/main/java/woowacourse/movie/booking/ConfirmDialog.kt
@@ -1,7 +1,8 @@
-package woowacourse.movie
+package woowacourse.movie.booking
 
 import android.app.AlertDialog
 import android.content.Context
+import woowacourse.movie.R
 import woowacourse.movie.dto.Ticket
 
 object ConfirmDialog {
