+++ app/src/main/java/woowacourse/movie/booking/MovieInfoGetter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.booking
 
 import java.time.DayOfWeek
 import java.time.LocalDate
